refactor(infrastructure): 포인트 금액 컬럼을 BIGINT 대신 DECIMAL로 저장
도메인의 금액 값 객체(PointAmount, Balance 등)는 모두 BigDecimal인데
JPA 엔티티는 Long/BIGINT였고, 매퍼에서 longValueExact/BigDecimal.valueOf로
방어적 변환을 거치며 타입 체계가 어긋나 있었다. 엔티티 필드를 BigDecimal로,
DB 컬럼을 DECIMAL(19,0)로 맞춰 변환 계층 없이 도메인과 영속 계층의 타입을
일치시킨다.

+++ infrastructure/src/main/kotlin/com/jysohn0825/point/infrastructure/persistence/adapter/PointWalletPersistenceAdapter.kt
@@ -29,7 +29,7 @@ class PointWalletPersistenceAdapter(
     ) {
         // holdingLimitOverride는 도메인 PointWallet이 들고 있지 않은 컬럼이므로(항상 balance만 변경),
         // 기존 row의 값을 그대로 보존해서 다시 쓴다. 신규 지갑이면 기존 row가 없어 null(정책값 적용)로 유지된다.
-        val existingOverride: Long? = jpaRepository.findById(wallet.id).orElse(null)?.holdingLimitOverride
+        val existingOverride: BigDecimal? = jpaRepository.findById(wallet.id).orElse(null)?.holdingLimitOverride
         jpaRepository.save(PointWalletMapper.of(wallet = wallet, memberId = memberId, holdingLimitOverride = existingOverride))
     }
 
@@ -56,6 +56,6 @@ class PointWalletPersistenceAdapter(
 
     private fun resolveHoldingLimit(entity: PointWalletEntity): HoldingLimit =
         entity.holdingLimitOverride
-            ?.let { HoldingLimit(BigDecimal(it)) }
+            ?.let { HoldingLimit(it) }
             ?: HoldingLimit(policyRepository.getCurrent().maxHoldingAmount.value)
 }
